+++ changelog.md
@@ -10,6 +10,7 @@
 * LLT-3948: Add IPv6 functionality and fallback to session_keeper module
 * LLT-3629: Add IPv6 support to nat-lab
 * LLT-4119: Fix multiple bugs in telio-firewall. Mainly whitelist vpn server
+* LLT-4166: Bump tokio-rustls version
 
 <br>